- moved all logic for auth from concern to base auth service - updated okta auth service - moved specs from common service concern to base auth service   specs - deleted common service concern and specs
